Add documentation for RepositoryCacheProxy & refactor its packages
