+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/data/repository_cache/InMemoryRepositoryCache.kt
@@ -1,5 +1,11 @@
 package org.hyperskill.app.core.data.repository_cache
 
+import org.hyperskill.app.core.domain.repository_cache.RepositoryCache
+
+/**
+ * Represent an in-memory cache based on [MutableMap].
+ * It is not safe to use it in a concurrent environment without additional synchronization.
+ */
 class InMemoryRepositoryCache<Key : Any, Value : Any?> : RepositoryCache<Key, Value> {
     private val cache: MutableMap<Key, Value> = mutableMapOf()
