--- shared/src/commonMain/kotlin/org/hyperskill/app/core/data/repository_cache/RepositoryCache.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.core.data.repository_cache
-
-interface RepositoryCache<Key : Any, Value : Any?> {
-    fun getAll(keys: List<Key>): List<Value>
-    fun putAll(map: Map<Key, Value>)
-    fun clearCache()
-}
