+++ shared/src/commonMain/kotlin/org/hyperskill/app/progresses/data/source/ProjectProgressesCacheDataSource.kt
@@ -1,6 +1,6 @@
 package org.hyperskill.app.progresses.data.source
 
-import org.hyperskill.app.core.data.repository_cache.RepositoryCache
+import org.hyperskill.app.core.domain.repository_cache.RepositoryCache
 import org.hyperskill.app.projects.domain.model.ProjectProgress
 
 interface ProjectProgressesCacheDataSource : RepositoryCache<Long, ProjectProgress>
