+++ shared/src/commonMain/kotlin/org/hyperskill/app/track/cache/TrackCacheDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package org.hyperskill.app.track.cache
 
 import org.hyperskill.app.core.data.repository_cache.InMemoryRepositoryCache
-import org.hyperskill.app.core.data.repository_cache.RepositoryCache
+import org.hyperskill.app.core.domain.repository_cache.RepositoryCache
 import org.hyperskill.app.track.data.source.TrackCacheDataSource
 import org.hyperskill.app.track.domain.model.Track
