+++ shared/src/commonMain/kotlin/org/hyperskill/app/track/data/source/TrackCacheDataSource.kt
@@ -1,6 +1,6 @@
 package org.hyperskill.app.track.data.source
 
-import org.hyperskill.app.core.data.repository_cache.RepositoryCache
+import org.hyperskill.app.core.domain.repository_cache.RepositoryCache
 import org.hyperskill.app.track.domain.model.Track
 
 interface TrackCacheDataSource : RepositoryCache<Long, Track>
